# 28demo/docs/swagger.yaml
basePath: /
definitions:
  main.Book:
    properties:
      author:
        type: string
      id:
        type: string
      stock:
        type: string
      title:
        type: string
    type: object
  main.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 实现对图书的增删改查的图书管理系统
  title: 图书管理系统
  version: "1.0"
paths:
  /books:
    get:
      description: 获取所有图书信息
      produces:
      - application/json
      responses:
        "200":
          description: 书籍信息
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/main.Book'
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有书籍
      tags:
      - 图书
    post:
      consumes:
      - application/json
      description: 传入书籍信息新增书籍
      parameters:
      - description: 书籍信息
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/main.Book'
      produces:
      - application/json
      responses:
        "201":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数解析失败
          schema:
            $ref: '#/definitions/main.Response'
        "409":
          description: 书籍ID已存在
          schema:
            $ref: '#/definitions/main.Response'
      security:
      - ApiKeyAuth: []
      summary: 添加书籍
      tags:
      - 图书
  /books/{id}:
    delete:
      description: 根据书籍ID删除书籍
      parameters:
      - description: 书籍ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 书籍删除成功
          schema:
            $ref: '#/definitions/main.Response'
        "404":
          description: 书籍不存在
          schema:
            $ref: '#/definitions/main.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除书籍
      tags:
      - 图书
    get:
      description: 获取指定ID的图书信息
      parameters:
      - description: 书籍ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 书籍信息
          schema:
            $ref: '#/definitions/main.Book'
        "404":
          description: 书籍不存在
          schema:
            $ref: '#/definitions/main.Response'
      security:
      - ApiKeyAuth: []
      summary: 根据ID查询书籍
      tags:
      - 图书
    put:
      consumes:
      - application/json
      description: 根据ID更新书籍信息
      parameters:
      - description: 书籍ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新后的书籍信息
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/main.Book'
      produces:
      - application/json
      responses:
        "200":
          description: 书籍更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数绑定失败
          schema:
            $ref: '#/definitions/main.Response'
        "404":
          description: 书籍不存在
          schema:
            $ref: '#/definitions/main.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新书籍
      tags:
      - 图书
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录并获取Token
      parameters:
      - description: 用户名和密码
        in: body
        name: user
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/main.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/main.Response'
      summary: 用户登录
      tags:
      - 用户
  /register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 用户名和密码
        in: body
        name: user
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/main.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/main.Response'
        "409":
          description: 用户已存在
          schema:
            $ref: '#/definitions/main.Response'
      summary: 用户注册
      tags:
      - 用户
swagger: "2.0"